feat: Add custom exceptions (#20)

+++ webapps/api-rest/src/main/java/ooo/alvar/nutrimenu/apirest/empresa/EmpresaService.java
@@ -1,5 +1,7 @@
 package ooo.alvar.nutrimenu.apirest.empresa;
 
+import ooo.alvar.nutrimenu.apirest.excepciones.EntityAlreadyExistsException;
+import ooo.alvar.nutrimenu.apirest.excepciones.EntityDoesntExistsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +14,14 @@ public class EmpresaService {
   @Autowired
   private EmpresaRepository empresaRepository;
 
+  public Empresa getEmpresa(String id) {
+    Empresa empresaDevuelta = empresaRepository.findById(id).orElse(null);
+    if (empresaDevuelta == null) {
+      throw new EntityDoesntExistsException("No existe una empresa con id " + id);
+    }
+    return empresaDevuelta;
+  }
+
   public List<Empresa> getAllEmpresas() {
     List<Empresa> empresas = new ArrayList<>();
     empresaRepository.findAll()
@@ -20,21 +30,30 @@ public class EmpresaService {
     return empresas;
   }
 
-  public Empresa getEmpresa(String id) {
-    return empresaRepository.findById(id).orElse(null);
-  }
-
-  public void addEmpresa(Empresa empresa) {
+  public Empresa addEmpresa(Empresa empresa) {
     empresa.setId(empresa.getNombre().toLowerCase().replace(' ', '_'));
-    empresaRepository.save(empresa);
+
+    if (empresaRepository.existsById(empresa.getId())) {
+      throw new EntityAlreadyExistsException("Ya existe una empresa llamada " + empresa.getNombre());
+    }
+
+    return empresaRepository.save(empresa);
   }
 
-  public void updateEmpresa(Empresa empresa, String id) {
+  public Empresa updateEmpresa(Empresa empresa, String id) {
+    if (!empresaRepository.existsById(id)) {
+      throw new EntityDoesntExistsException("No existe una empresa con id " + id);
+    }
+
     empresa.setId(id);
-    empresaRepository.save(empresa);
+    return empresaRepository.save(empresa);
   }
 
   public void deleteEmpresa(String id) {
+    if (!empresaRepository.existsById(id)) {
+      throw new EntityDoesntExistsException("No existe una empresa con id " + id);
+    }
+
     empresaRepository.deleteById(id);
   }
 }
