+++ webapps/api-rest/src/main/java/ooo/alvar/nutrimenu/apirest/menu/MenuService.java
@@ -1,5 +1,6 @@
 package ooo.alvar.nutrimenu.apirest.menu;
 
+import ooo.alvar.nutrimenu.apirest.excepciones.EntityDoesntExistsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +13,14 @@ public class MenuService {
   @Autowired
   private MenuRepository menuRepository;
 
+  public Menu getMenu(String id) {
+    Menu menuDevuelto = menuRepository.findById(id).orElse(null);
+    if (menuDevuelto == null) {
+      throw new EntityDoesntExistsException("No existe un menú con id " + id);
+    }
+    return menuDevuelto;
+  }
+
   public List<Menu> getAllMenusByEmpresa(String id) {
     List<Menu> menus = new ArrayList<>();
     menuRepository.findByEmpresaId(id)
@@ -20,11 +29,7 @@ public class MenuService {
     return menus;
   }
 
-  public Menu getMenu(String id) {
-    return menuRepository.findById(id).orElse(null);
-  }
-
-  public void addMenu(Menu menu) {
+  public Menu addMenu(Menu menu) {
     Menu ultimoMenu = menuRepository.findTopByIdContainsIgnoreCaseOrderByIdDesc(menu.getEmpresa().getId() + "-" + menu.getNombre().replace(' ', '_')).orElse(null);
     int contador = 1;
 
@@ -38,18 +43,26 @@ public class MenuService {
 
     menu.setFechaCreacion(java.time.Instant.now());
     menu.setFechaModificacion(java.time.Instant.now());
-    menuRepository.save(menu);
+    return menuRepository.save(menu);
   }
 
-  public void updateMenu(Menu menu, String id) {
+  public Menu updateMenu(Menu menu, String id) {
+    if (!menuRepository.existsById(id)) {
+      throw new EntityDoesntExistsException("No existe un menú con id " + id);
+    }
+
     menu.setId(id);
     Menu menuOriginal = menuRepository.findById(id).orElse(null);
     menu.setFechaCreacion(menuOriginal.getFechaCreacion());
     menu.setFechaModificacion(java.time.Instant.now());
-    menuRepository.save(menu);
+    return menuRepository.save(menu);
   }
 
   public void deleteMenu(String id) {
+    if (!menuRepository.existsById(id)) {
+      throw new EntityDoesntExistsException("No existe un menú con id " + id);
+    }
+
     menuRepository.deleteById(id);
   }
 }
